fix to actually upload. Duh!

+++ xwords4/android/scripts/build-debug.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 
+set -e -u
+
+NO_RM=''
+
 DIR=/tmp/build_$$_dir
 mkdir -p $DIR
 pushd $DIR
 
 usage() {
     [ $# -ge 1 ] && echo "Error: $1"
+	echo "usage: $0 [--no-rm]    # do not remove the build directory"
     echo "builds debug variant from the current tip of github"
 	echo "(last modified Jan 2024)"
     exit 1
@@ -15,6 +20,9 @@ while [ $# -ge 1 ]; do
     case $1 in
 		--help)
 			usage
+			;;
+		--no-rm)
+			NO_RM=1
 			;;
 		*)
 			usage "unexpected command $1"
@@ -31,7 +39,7 @@ APK="$(find . -name '*.apk')"
 if [ -n "${XW4D_UPLOAD}" ]; then
 	IFS=","
 	for UPPATH in ${XW4D_UPLOAD}; do
-		# scp "$APK" ${UPPATH}
+		scp "$APK" ${UPPATH}
 		echo "uploaded $APK to ${UPPATH}"
 	done
 else
@@ -39,4 +47,8 @@ else
 fi
 
 popd
-rm -rf $DIR
+if [ -s "${NO_RM}" ]; then
+	rm -rf $DIR
+else
+	echo "not removing: $DIR"
+fi
